04.12.2018

+++ app/Http/Controllers/Admin/Api/ProductController.php
@@ -11,7 +11,9 @@ use App\Models\AttributeSet;
 use App\Models\ProductGroup;
 use App\Models\ProductImage;
 use App\Models\SpecificPrice;
+use App\Requests\CloneProductRequest;
 use App\Requests\SaveProductRequest;
+use App\Services\ServiceAttributeProductValue;
 use App\Tools\UploadableTrait;
 use Illuminate\Http\Request;
 use DB;
@@ -20,14 +22,20 @@ use File;
 class ProductController extends AdminController
 {
     use UploadableTrait;
+    private $serviceAttrProductVal;
 
+    public function __construct(ServiceAttributeProductValue $serviceAttrProductVal)
+    {
+        parent::__construct();
+        $this->serviceAttrProductVal = $serviceAttrProductVal;
+    }
 
     public function list(Request $request)
     {
 
         $search = trim(mb_strtolower($request->input('search')));
 
-        $list =  Product::with([
+        $list =  Product::withTrashed()->with([
                 'categories' => function($query){
                     $query->select([DB::raw('t_categories.id'), DB::raw('t_categories.name')]);
                 }
@@ -51,10 +59,13 @@ class ProductController extends AdminController
         return  $this->sendResponse($list);
     }
 
-    public function save(SaveProductRequest $req){
+    public function save(SaveProductRequest $req)
+    {
 
         $request = $req->all();
+
         $reqProduct = $request['product'];
+        $reqProduct['deleted_at'] = intval($reqProduct['deleted_at']) == 1 ? date('Y-m-d H:i:s') : NULL;
 
         if(empty($reqProduct['id']))
         {
@@ -62,19 +73,26 @@ class ProductController extends AdminController
             $reqProduct['group_id'] = $productGroup->id;
         }
 
-        $product = Product::findOrNew($reqProduct['id']);
+        $product = Product::withTrashed()->findOrNew($reqProduct['id']);
+
+
+
+
+
+
         $old_attribute_set_id = $product->attribute_set_id;
+
+
 
         $product->fill($reqProduct);
 
         if($product->save())
         {
 
-            if($req->file("product_photo"))
+            if($req->file("photo"))
             {
                 $product->deletePhoto();
-                $photo = $this->uploadFile($request['product_photo'], config('shop.products_path_file') . $product->id . '/');
-                $product->photo = $photo;
+                $product->photo = $this->uploadFile($request['photo'], $product->productFileFolder());
                 $product->save();
             }
 
@@ -83,15 +101,18 @@ class ProductController extends AdminController
 
 
             //Конкретная цена
-            $SpecificPrice = SpecificPrice::firstOrNew(['product_id' => $product->id]);
-            $SpecificPrice->fill($request['specific_price']);
-            $SpecificPrice->save();
+            if(!empty($request['specific_price']['reduction']))
+            {
+                $SpecificPrice = SpecificPrice::firstOrNew(['product_id' => $product->id]);
+                $SpecificPrice->fill($request['specific_price']);
+                $SpecificPrice->save();
+            }
 
 
             //Атрибуты
 
             // Get old product atrribute values
-            $oldAttributes = $product->attributeProductValue;
+            $oldAttributes = $product->attributes;
 
             if(!empty($reqProduct['id']))
                     $product->attributes()->detach();
@@ -100,10 +121,7 @@ class ProductController extends AdminController
             {
                 foreach ($oldAttributes as $oldAttr)
                 {
-                    if(File::exists(config('shop.attributes_path_file') . $oldAttr->value)
-                        and
-                        !$product->AttributeValue::where('attribute_id', $oldAttr->attribute_id)->where('value', $oldAttr->value)->first())
-                            File::delete(config('shop.attributes_path_file') . $oldAttr->value);
+                    $this->serviceAttrProductVal->deleteImage($oldAttr->pivot->attribute_id, $oldAttr->pivot->value);
                 }
             }
 
@@ -118,10 +136,9 @@ class ProductController extends AdminController
                     {
                         foreach ($oldAttributes as $oldAttr)
                         {
-                            if($oldAttr->attribute_id == $item['attribute_id'])
+                            if($oldAttr->pivot->attribute_id == $item['attribute_id'])
                             {
-                                if(!AttributeValue::where('attribute_id', $oldAttr->attribute_id)->where('value', $oldAttr->value)->first())
-                                    File::delete(config('shop.attributes_path_file') . $oldAttr->value);
+                                $this->serviceAttrProductVal->deleteImage($oldAttr->pivot->attribute_id, $oldAttr->pivot->value);
                             }
                         }
                     }
@@ -149,7 +166,7 @@ class ProductController extends AdminController
                         if($req->file("product_images.$key.value"))
                             $images = [
                                 'product_id' => $product->id,
-                                'name' => $this->uploadFile($item['value'], config('shop.products_path_file') . $product->id . '/'),
+                                'name' => $this->uploadFile($item['value'], $product->productFileFolder()),
                                 'order' => $key
                             ];
                         else
@@ -167,26 +184,29 @@ class ProductController extends AdminController
 
 
             //Группа товаров
-            $products = Product::select('id')->where('group_id', $product->group_id)->whereNotIn('id', $request['products_ids_group'])->get();
-            foreach ($products as $item)
+            if(is_array($request['products_ids_group']))
             {
-                if($item->id == $product->id) continue;
+                $products = Product::withTrashed()->select('id')->where('group_id', $product->group_id)->whereNotIn('id', $request['products_ids_group'])->get();
+                foreach ($products as $item)
+                {
+                    if($item->id == $product->id) continue;
 
-                $productGroup = ProductGroup::create();
-                Product::where('id', $item->id)->update(['group_id' => $productGroup->id]);
-            }
+                    $productGroup = ProductGroup::create();
+                    Product::withTrashed()->where('id', $item->id)->update(['group_id' => $productGroup->id]);
+                }
 
-            foreach ($request['products_ids_group'] as $products_id)
-            {
-                if($products_id == $product->id) continue;
+                foreach ($request['products_ids_group'] as $products_id)
+                {
+                    if($products_id == $product->id) continue;
 
-                $productGroupId     = Product::find($products_id)->group_id;
-                $countGroupProducts = Product::where('group_id', $productGroupId)->count();
+                    $productGroupId     = Product::withTrashed()->find($products_id)->group_id;
+                    $countGroupProducts = Product::withTrashed()->where('group_id', $productGroupId)->count();
 
-                Product::where('id', $products_id)->update(['group_id' => $product->group_id]);
+                    Product::withTrashed()->where('id', $products_id)->update(['group_id' => $product->group_id]);
 
-                if($countGroupProducts == 1)
-                    ProductGroup::destroy($productGroupId);
+                    if($countGroupProducts == 1)
+                        ProductGroup::destroy($productGroupId);
+                }
             }
 
         }
@@ -195,7 +215,7 @@ class ProductController extends AdminController
 
     public function view($id)
     {
-        $product = Product::with(['categories', 'attributeProductValue', 'specificPrice', 'images'])->findOrFail($id);
+        $product = Product::withTrashed()->with(['categories', 'attributes', 'specificPrice', 'images'])->findOrFail($id);
 
         //категория
         $categories = $product->categories->pluck('id');
@@ -205,10 +225,10 @@ class ProductController extends AdminController
 
         //атрибуты
         $data = [];
-        foreach ($product->attributeProductValue as $item)
-            $data[$item->attribute_id][] = $item->value;
-        unset($product->attributeProductValue);
-        $product->attribute_product_value = $data;
+        foreach ($product->attributes as $item)
+            $data[$item->pivot->attribute_id][] = $item->pivot->value;
+        unset($product->attributes);
+        $product->attributes = $data;
 
         
         //картинки
@@ -216,7 +236,7 @@ class ProductController extends AdminController
             return  [
                 'id'         => $item->id,
                 'is_delete'  => 0,
-                'image_view' => $item->imagePath(),
+                'image_view' => $item->imagePath(true),
                 'value'      => ''
             ];
         });
@@ -225,7 +245,7 @@ class ProductController extends AdminController
 
 
         //фото товара
-        $product->product_photo = $product->pathPhoto();
+        $product->photo = $product->pathPhoto(true);
 
 
         return  $this->sendResponse($product);
@@ -233,9 +253,119 @@ class ProductController extends AdminController
 
     public function groupProducts($group_id)
     {
-        $products = Product::where('group_id', $group_id)->get();
+        $products = Product::withTrashed()->where('group_id', $group_id)->get();
         return  $this->sendResponse($products);
     }
+
+    public function delete($id)
+    {
+
+        $product = Product::withTrashed()->find($id);
+
+        //категории
+        $product->categories()->detach();
+
+        //картинки
+        if($product->images)
+            $product->images()->delete();
+
+        //арритуты
+        if($product->attributes)
+        {
+            foreach ($product->attributes as $item)
+                if($item->type == 'media')
+                    $this->serviceAttrProductVal->deleteImage($item->pivot->attribute_id, $item->pivot->value);
+            $product->attributes()->detach();
+        }
+
+        //папка товара
+        File::deleteDirectory($product->productFileFolder());
+
+        //скидки
+        if($product->specificPrice)
+            $product->specificPrice->delete();
+
+
+        $group_id = $product->group_id;
+
+        if($product->delete())
+        {
+            $countGroupProducts = Product::withTrashed()->where('group_id', $group_id)->count();
+            if(!$countGroupProducts)
+                ProductGroup::destroy($group_id);
+
+            return  $this->sendResponse(true);
+        }
+        return  $this->sendResponse(false);
+    }
+
+
+    public function cloneProduct(CloneProductRequest $req)
+    {
+        $req = $req->input('clone_product');
+
+        $product = Product::withTrashed()->find($req['product_id']);
+
+        // Create clone object
+        $clone = $product->replicate();
+
+        //Группа товаров
+        if(!$req['group'])
+            $clone->group_id = ProductGroup::create()->id;
+
+        $clone->sku  = $req['sku'];
+        $clone->name = $req['name'];
+
+        //Save cloned product
+        $clone->push();
+
+        //Фото товара
+        if($req['photo'] and $product->pathPhoto())
+        {
+            $clone->photo = $this->copyFile($product->pathPhoto(), $clone->productFileFolder());
+            $clone->save();
+        }
+
+
+        //категория
+        foreach ($product->categories as $item)
+            $clone->categories()->attach([$item->id]);
+
+        //атрибуты
+        if($req['attributes'])
+        {
+            foreach ($product->attributes as $item)
+                $clone->attributes()->attach([$item->pivot->attribute_id => ['value' =>  $item->pivot->value]]);
+        }
+
+        //Скидки
+        if($req['specific_price'])
+        {
+            $specificPrice = $product->specificPrice->replicate();
+            $specificPrice->product_id = $clone->id;
+            $specificPrice->push();
+        }
+
+        //Картинки
+        if($req['product_images'])
+        {
+            $images = $product->images;
+            if($images)
+            {
+                foreach ($images as $image)
+                {
+                    $data = $image->toArray();
+                    $data['product_id'] = $product->id;
+                    $data['name'] = $this->copyFile($image->imagePath(), $clone->productFileFolder());
+                    $clone->images()->create($data);
+                }
+            }
+        }
+
+
+        return  $this->sendResponse(true);
+    }
+
